+++ Assets/00.WorkSpace/GIL/Scripts/Managers/BlockSpawnManager.cs
@@ -1,19 +1,10 @@
 using System.Collections.Generic;
-using _00.WorkSpace.GIL.Scripts.Grids;
 using _00.WorkSpace.GIL.Scripts.Shapes;
 using UnityEngine;
-using Debug = UnityEngine.Debug;
-using Random = UnityEngine.Random;
 
 namespace _00.WorkSpace.GIL.Scripts.Managers
 {
-    public struct FitInfo
-    {
-        public Vector2Int Offset;                 // 좌상단 오프셋 (col=x, row=y)
-        public List<GridSquare> CoveredSquares;   // 이 배치로 덮게 될 셀들
-    }
-    
-    public class BlockSpawnManager : MonoBehaviour
+    public partial class BlockSpawnManager : MonoBehaviour
     {
         public static BlockSpawnManager Instance { get; private set; }
 
@@ -28,6 +19,10 @@ namespace _00.WorkSpace.GIL.Scripts.Managers
 
         [Header("Wave Rules")] 
         [SerializeField, Range(1, 3)] private int maxDuplicatesPerWave = 2;
+        
+        [Header("Wave Record Rules")]
+        [SerializeField, Range(2, 5)] private int maxSameWaveStreak = 3;
+        private readonly Queue<string> _lastWaves = new();
         
         private int[] _cumulativeWeights;
         private int[] _inverseCumulativeWeights;
@@ -62,315 +57,16 @@ namespace _00.WorkSpace.GIL.Scripts.Managers
             BuildInverseCumulativeTable();
         }
         
-        /// <summary>
-        /// 가중치 기반으로 블럭 생성, 3개 이하 블럭에 대해서는 별도의 생성 확률 적용
-        /// </summary>
-        public List<ShapeData> GenerateBasicWave(int count)
-        {
-            var result = new List<ShapeData>(count);
-            
-            // 이번 웨이브에서 "소환 실패한 블록" 은 이후 검색에서 제외
-            var excludedByPenalty = new HashSet<string>();
-            var excludedByDupes = new HashSet<string>();
-            var perShapeCount = new Dictionary<string, int>();
-            
-            for (int i = 0; i < count; i++)
-            {
-                ShapeData chosen = null;
-                int guard = 0; // 무한 루프 방지
-
-                while (chosen == null && guard++ < shapeData.Count)
-                {
-                    // 제외 목록 반영해서 가중치 추첨
-                    var pick = GetRandomShapeByWeightExcluding(excludedByPenalty, excludedByDupes);
-                    if (pick == null) break; // 전부 제외된 경우
-
-                    // 소형 여부를 activeBlockCount로 판정
-                    bool isSmall = pick.activeBlockCount <= smallBlockTileThreshold;
-                    if (smallBlockPenaltyMode && isSmall && Random.value < smallBlockFailRate)
-                    {
-                        // 소환 실패 → 이번 웨이브에서 제외하고 재검색
-                        excludedByPenalty.Add(pick.Id);
-                        continue;
-                    }
-
-                    // 성공
-                    chosen = pick;
-                }
-                // 소형 패널티는 무시하지만, 중복 한도 제외는 반드시 시킴
-                if (chosen == null)
-                {
-                    chosen = GetRandomShapeByWeightExcluding(null, excludedByDupes);
-                    // 그래도 못 뽑았으면 가중치 기반 생성 ( 4개 이상 블럭은 삭제되지 않아서 발생할 확률이 없음 )
-                    if (chosen == null)
-                    {
-                        Debug.LogWarning("모든 후보가 소형 페널티로 제외되어 가중치 강제 소환을 수행합니다.");                    
-                        chosen = GetRandomShapeByWeight();
-                    }
-                }
-                result.Add(chosen);
-                
-                string chosenId = chosen.Id;
-                perShapeCount.TryGetValue(chosenId, out int cnt);
-                cnt++;
-                perShapeCount[chosenId] = cnt;
-                if (cnt >= maxDuplicatesPerWave)
-                {
 #if UNITY_EDITOR
-                    Debug.Log($"{maxDuplicatesPerWave}이상 중복됨, 다음 선택에서 제외");
-#endif                    
-                    excludedByDupes.Add(chosenId); // 다음 선택에서 제외.
-                }
-            }
-
-            return result;
-        }
-        
-        private ShapeData GetRandomShapeByWeightExcluding(HashSet<string> exPenalty, HashSet<string> exDupes)
+        private void OnValidate()
         {
-            if (shapeData == null || shapeData.Count == 0) return null;
-
-            var total = 0;
-            for (int i = 0; i < shapeData.Count; i++)
+            if (!Application.isPlaying)
             {
-                var s = shapeData[i];
-                if ((exPenalty != null && exPenalty.Contains(s.Id)) ||
-                    (exDupes   != null && exDupes.Contains(s.Id))) continue;
-                total += s.chanceForSpawn;
-            }
-            if (total <= 0) return null;
-
-            var r = Random.Range(0, total); // [0,total)
-            var acc = 0;
-            for (int i = 0; i < shapeData.Count; i++)
-            {
-                var s = shapeData[i];
-                if ((exPenalty != null && exPenalty.Contains(s.Id)) ||
-                    (exDupes   != null && exDupes.Contains(s.Id))) continue;
-                acc += s.chanceForSpawn;
-                if (r < acc) return s;
-            }
-            return null;
-        }
-        
-        private void BuildCumulativeTable()
-        {
-            _cumulativeWeights = new int[shapeData.Count];
-            _totalWeight = 0;
-
-            for (int i = 0; i < shapeData.Count; i++)
-            {
-                _totalWeight += shapeData[i].chanceForSpawn;
-                _cumulativeWeights[i] = _totalWeight;
-            }
-
-            Debug.Log($"가중치 계산 완료: {_totalWeight}");
-        }
-
-        private void BuildInverseCumulativeTable()
-        {
-            _inverseCumulativeWeights = new int[shapeData.Count];
-            _inverseTotalWeight = 0;
-
-            for (int i = 0; i < shapeData.Count; i++)
-            {
-                _inverseTotalWeight += (_totalWeight - shapeData[i].chanceForSpawn);
-                _inverseCumulativeWeights[i] = _inverseTotalWeight;
-            }
-
-            Debug.Log($"역가중치 계산 완료: {_inverseTotalWeight}");
-        }    
-        
-        private ShapeData GetRandomShapeByWeight()
-        {
-            if (shapeData == null || shapeData.Count == 0) return null;
-            if (_cumulativeWeights == null || _cumulativeWeights.Length != shapeData.Count) BuildCumulativeTable();
-
-            int r = Random.Range(0, Mathf.Max(1, _totalWeight));
-            for (int i = 0; i < _cumulativeWeights?.Length; i++)
-            {
-                if (r < _cumulativeWeights[i])
-                {
-                    return shapeData[i];
-                }
-            }
-            
-            return shapeData[^1];
-        }
-        
-        public bool CanPlaceShapeData(ShapeData shape)
-        {
-            var gm = GridManager.Instance;
-            var states = gm.gridStates;
-
-            var (minX, maxX, minY, maxY) = GetShapeBounds(shape);
-            int shapeRows = maxY - minY + 1;
-            int shapeCols = maxX - minX + 1;
-
-            for (int yOff = 0; yOff <= gm.rows - shapeRows; yOff++)
-            {
-                for (int xOff = 0; xOff <= gm.cols - shapeCols; xOff++)
-                {
-                    if (CanPlace(shape, minX, minY, shapeRows, shapeCols, states, yOff, xOff))
-                        return true;
-                }
-            }
-
-            return false;
-        }
-
-        private bool CanPlace(ShapeData shape, int minX, int minY, int shapeRows, int shapeCols,
-            bool[,] grid, int yOffset, int xOffset)
-        {
-            for (int y = 0; y < shapeRows; y++)
-            for (int x = 0; x < shapeCols; x++)
-            {
-                if (!shape.rows[y + minY].columns[x + minX]) continue;
-                if (grid[y + yOffset, x + xOffset]) return false;
-            }
-
-            return true;
-        }
-
-        private static (int minX, int maxX, int minY, int maxY) GetShapeBounds(ShapeData shape)
-        {
-            int minX = int.MaxValue, minY = int.MaxValue;
-            int maxX = int.MinValue, maxY = int.MinValue;
-
-            for (int y = 0; y < shape.rows.Length; y++)
-            {
-                for (int x = 0; x < shape.rows[y].columns.Length; x++)
-                {
-                    if (!shape.rows[y].columns[x]) continue;
-                    if (x < minX) minX = x;
-                    if (x > maxX) maxX = x;
-                    if (y < minY) minY = y;
-                    if (y > maxY) maxY = y;
-                }
-            }
-
-            if (maxX < 0)
-            {
-                minX = minY = 0;
-                maxX = maxY = 0;
-            }
-
-            return (minX, maxX, minY, maxY);
-        }
-
-        /// <summary>
-        /// 현 보드(실제 상태 기준)에서 shape를 놓을 수 있는 모든 위치 중
-        /// '겹치지 않는' 하나를 찾아 반환 (좌상단 우선). 못 찾으면 false.
-        /// </summary>
-        public bool TryFindOneFit(ShapeData shape, bool[,] virtualBoard, out FitInfo fit)
-        {
-            fit = default;
-            var gm = GridManager.Instance;
-            var squares = gm.gridSquares;
-
-            // 경계 상자
-            var (minX, maxX, minY, maxY) = GetShapeBounds(shape);
-            int shRows = maxY - minY + 1, shCols = maxX - minX + 1;
-
-            for (int oy = 0; oy <= gm.rows - shRows; oy++)
-            {
-                for (int ox = 0; ox <= gm.cols - shCols; ox++)
-                {
-                    bool ok = true;
-                    var list = new List<GridSquare>(8);
-
-                    for (int y = 0; y < shRows && ok; y++)
-                    for (int x = 0; x < shCols; x++)
-                    {
-                        if (!shape.rows[y + minY].columns[x + minX]) continue;
-
-                        // 실제/가상 보드 중 가상 보드 우선(중복 방지)
-                        bool occupied = virtualBoard != null
-                            ? virtualBoard[oy + y, ox + x]
-                            : squares[oy + y, ox + x].IsOccupied;
-
-                        if (occupied) { ok = false; break; }
-                        list.Add(squares[oy + y, ox + x]);
-                    }
-
-                    if (ok)
-                    {
-                        fit = new FitInfo { Offset = new Vector2Int(ox, oy), CoveredSquares = list };
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-
-        
-
-        /// <summary>
-        /// 웨이브 전체에 대해 겹치지 않는 위치를 계산하고 Hover로 표시.
-        /// 같은 셀 중복 하이라이트를 막기 위해 가상보드에 순차 점유 마킹.
-        /// </summary>
-        public void PreviewWaveNonOverlapping(List<ShapeData> wave, List<Sprite> spritesOrNull)
-        {
-            ClearPreview();
-
-            var gm = GridManager.Instance;
-            var rows = gm.rows; 
-            var cols = gm.cols;
-            var squares = gm.gridSquares;
-
-            // 가상 보드: 현재 점유 상태를 복사하고, 미리보기로 선점한 칸은 true로 마킹
-            var virtualBoard = new bool[rows, cols];
-            for (int row = 0; row < rows; row++)
-                for (int col = 0; col < cols; col++)
-                    virtualBoard[row, col] = squares[row, col].IsOccupied;
-
-            for (int i = 0; i < wave.Count; i++)
-            {
-                if (wave[i] == null) continue;
-                if (TryFindOneFit(wave[i], virtualBoard, out var fit))
-                {
-                    // 스프라이트는 선택 사항: null이면 그리드의 기본 hover이미지로 표시됨
-                    var sprite = (spritesOrNull != null && i < spritesOrNull.Count) ? spritesOrNull[i] : null;
-                    ApplyPreview(fit, sprite);
-
-                    // 가상보드 점유 마킹(다음 블록 프리뷰가 겹치지 않게)
-                    foreach (var sq in fit.CoveredSquares)
-                        virtualBoard[sq.RowIndex, sq.ColIndex] = true;
-                }
+                if (shapeData == null || shapeData.Count == 0)
+                    LoadResources();
+                BuildWeightTable();
             }
         }
-
-        /// <summary>
-        /// 지정된 배치의 셀들을 Hover 상태로 표시(점유 갱신 X)
-        /// </summary>
-        private void ApplyPreview(FitInfo fit, Sprite spriteOrNull)
-        {
-            foreach (var sq in fit.CoveredSquares)
-            {
-                if (spriteOrNull != null) 
-                    sq.SetImage(spriteOrNull);
-                if (!sq.IsOccupied) 
-                    sq.SetState(GridState.Hover); // Hover 이미지를 켬
-            }
-        }
-
-        /// <summary>
-        /// 현재 보드에서 점유되지 않은 셀들의 Hover를 모두 해제
-        /// </summary>
-        public void ClearPreview()
-        {
-            var gm = GridManager.Instance;
-            var squares = gm.gridSquares;
-            for (int r = 0; r < gm.rows; r++)
-            {
-                for (int c = 0; c < gm.cols; c++)
-                {
-                    if (!squares[r, c].IsOccupied)
-                        squares[r, c].SetState(GridState.Normal);
-                }
-            }
-        }
+#endif
     }
 }
-
-
