+++ Assets/00.WorkSpace/GIL/Scripts/Managers/BlockSpawnManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 40f7f7b8903350d4888c370f7f7177a9
+guid: 5030e7cc3244c2d47ad6a0e566687465
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
